refactor: Implement stubs, add Haddock docs, improve code quality
Implemented stubs:
- DB/HRCharge: Implement salaryChargeRow decoder using SalaryCharge type
- Surypus/Metrics/Prometheus: Implement insert using Map.insert
- Core/Event: Implement logEvent with getCurrentTime
- Core/Accounting: Implement processTransaction (validates and returns success)
- DAL/Mutations: Implement createUser/updateUser/authenticateUser with UserInput
- DAL/Types: Add UserInput type for user operations

Documentation improvements:
- Surypus/Types: Comprehensive module docs for Decimal type
- Surypus/Config: Full Haddock for all functions and types
- Surypus/Validation: Module-level and function docs
- Core/Accounting/Operations: Extensive invariant documentation
- Core/Tax: Already has LiquidHaskell annotations

Code improvements:
- Config.hs: Streamlined service container, reduced boilerplate
- Validation.hs: Better organized with documented validators
- Database/Pool.hs: Added validatePoolConnection for health checks

Test improvements:
- Added QuickCheck properties for VAT and double-entry bookkeeping
- All existing tests maintained

+++ src/Core/Payroll/Types.hs
@@ -178,4 +178,4 @@ totalAccruals = sum . filter (> 0) . fmap sAmount
 -- ============================================================================
 
 diffDays :: Day -> Day -> Int
-diffDays a b = fromIntegral (fromEnum a - fromEnum b)
+diffDays a b = fromEnum a - fromEnum b
